Update TokenData field types to i64

+++ photon-api/src/models/token_data.rs
@@ -13,12 +13,12 @@ use crate::models;
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct TokenData {
     #[serde(rename = "amount")]
-    pub amount: i32,
+    pub amount: i64,
     /// A Solana public key represented as a base58 string.
     #[serde(rename = "delegate", skip_serializing_if = "Option::is_none")]
     pub delegate: Option<String>,
     #[serde(rename = "isNative", skip_serializing_if = "Option::is_none")]
-    pub is_native: Option<i32>,
+    pub is_native: Option<i64>,
     /// A Solana public key represented as a base58 string.
     #[serde(rename = "mint")]
     pub mint: String,
@@ -30,7 +30,7 @@ pub struct TokenData {
 }
 
 impl TokenData {
-    pub fn new(amount: i32, mint: String, owner: String, state: models::AccountState) -> TokenData {
+    pub fn new(amount: i64, mint: String, owner: String, state: models::AccountState) -> TokenData {
         TokenData {
             amount,
             delegate: None,
